+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [v2.7.0] (2024-07-03)
+
+### Added
+
+- (react-native) Support remote debugging in Chrome [#468](https://github.com/bugsnag/bugsnag-js-performance/pull/468)
+
+### Fixed
+
+- (plugin-react-navigation) Fix a crash when navigation ref is created using useRef [#469](https://github.com/bugsnag/bugsnag-js-performance/pull/469)
+
 ## [v2.6.0] (2024-06-06)
 
 ### Added